moved location of variables storing providers api url and names to their respective provider classes
